Revert "Merge the e2e tests"
This reverts commit e6ab4e71ee94287934edf8ebe19bbbf1cf182df0.
